requirements.txt
annotated-types==0.7.0
anthropic==0.83.0
antlr4-python3-runtime==4.11.1
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.10.5
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
google-genai==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.31.0
ipython==8.18.1
jedi==0.19.2
jiter==0.11.1
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.2.1
mcp==1.26.0
mpmath==1.3.0
nest-asyncio==1.6.0
openai==2.7.1
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pyzmq==27.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tomli==2.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
wcwidth==0.2.14
zipp==3.23.0
